Seperate the definitions from the sync so the seed can require the definitions This means we only have to change db.js when we want to modify the tables, the changes will happen to db_config and db_seed
